refractor: moves subs tests in integration proj
