+++ Core.sln
@@ -14,25 +14,23 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleApp", "Examples\ExampleApp.csproj", "{73BBCC3C-E649-4977-A34A-7CF424D638E6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExampleApp", "Examples\ExampleApp.csproj", "{73BBCC3C-E649-4977-A34A-7CF424D638E6}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "GraphQLClient", "GraphQLClient", "{1099033B-8BD9-4038-AF9A-94A3370B4F8C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.Client", "GraphQLClient\src\GraphQL.Client\GraphQL.Client.csproj", "{390B0BA1-4621-459F-8CE2-788797AB3134}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.Client", "GraphQLClient\src\GraphQL.Client\GraphQL.Client.csproj", "{390B0BA1-4621-459F-8CE2-788797AB3134}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.Primitives", "GraphQLClient\src\GraphQL.Primitives\GraphQL.Primitives.csproj", "{A71D3F8F-4592-4CD4-83E4-A215D116C730}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.Primitives", "GraphQLClient\src\GraphQL.Primitives\GraphQL.Primitives.csproj", "{A71D3F8F-4592-4CD4-83E4-A215D116C730}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.Client.Abstractions", "GraphQLClient\src\GraphQL.Client.Abstractions\GraphQL.Client.Abstractions.csproj", "{87A3348F-A510-4101-BD87-DBBF9EBE4070}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.Client.Abstractions", "GraphQLClient\src\GraphQL.Client.Abstractions\GraphQL.Client.Abstractions.csproj", "{87A3348F-A510-4101-BD87-DBBF9EBE4070}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.Client.Abstractions.Websocket", "GraphQLClient\src\GraphQL.Client.Abstractions.Websocket\GraphQL.Client.Abstractions.Websocket.csproj", "{9C491E50-B0F8-4D08-BFAC-E0A2AD6CF563}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.Client.Abstractions.Websocket", "GraphQLClient\src\GraphQL.Client.Abstractions.Websocket\GraphQL.Client.Abstractions.Websocket.csproj", "{9C491E50-B0F8-4D08-BFAC-E0A2AD6CF563}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.Client.Serializer.Newtonsoft", "GraphQLClient\src\GraphQL.Client.Serializer.Newtonsoft\GraphQL.Client.Serializer.Newtonsoft.csproj", "{54B5BF04-3C46-498C-AE2A-1F2F85768C2C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.Client.Serializer.Newtonsoft", "GraphQLClient\src\GraphQL.Client.Serializer.Newtonsoft\GraphQL.Client.Serializer.Newtonsoft.csproj", "{54B5BF04-3C46-498C-AE2A-1F2F85768C2C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.Client.Serializer.SystemTextJson", "GraphQLClient\src\GraphQL.Client.Serializer.SystemTextJson\GraphQL.Client.Serializer.SystemTextJson.csproj", "{EAC54DBF-B8FC-4B20-A33A-3EBA05841128}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.Client.Serializer.SystemTextJson", "GraphQLClient\src\GraphQL.Client.Serializer.SystemTextJson\GraphQL.Client.Serializer.SystemTextJson.csproj", "{EAC54DBF-B8FC-4B20-A33A-3EBA05841128}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.Client.LocalExecution", "GraphQLClient\src\GraphQL.Client.LocalExecution\GraphQL.Client.LocalExecution.csproj", "{6A3861F4-CA3E-4790-BF27-3C502E18783C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestsSubscriptions", "TestsSubscriptions\TestsSubscriptions.csproj", "{33044D53-33E4-490B-BFD9-FB3F5B731B5F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.Client.LocalExecution", "GraphQLClient\src\GraphQL.Client.LocalExecution\GraphQL.Client.LocalExecution.csproj", "{6A3861F4-CA3E-4790-BF27-3C502E18783C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -84,22 +82,9 @@ Global
 		{6A3861F4-CA3E-4790-BF27-3C502E18783C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6A3861F4-CA3E-4790-BF27-3C502E18783C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6A3861F4-CA3E-4790-BF27-3C502E18783C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{33044D53-33E4-490B-BFD9-FB3F5B731B5F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{33044D53-33E4-490B-BFD9-FB3F5B731B5F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{33044D53-33E4-490B-BFD9-FB3F5B731B5F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{33044D53-33E4-490B-BFD9-FB3F5B731B5F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1378F8F4-B354-4DB6-8052-A1A69B59BC2C}
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		Policies = $0
-		$0.DotNetNamingPolicy = $1
-		$1.DirectoryNamespaceAssociation = PrefixedHierarchical
-		$0.TextStylePolicy = $4
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{390B0BA1-4621-459F-8CE2-788797AB3134} = {1099033B-8BD9-4038-AF9A-94A3370B4F8C}
@@ -109,5 +94,14 @@ Global
 		{54B5BF04-3C46-498C-AE2A-1F2F85768C2C} = {1099033B-8BD9-4038-AF9A-94A3370B4F8C}
 		{EAC54DBF-B8FC-4B20-A33A-3EBA05841128} = {1099033B-8BD9-4038-AF9A-94A3370B4F8C}
 		{6A3861F4-CA3E-4790-BF27-3C502E18783C} = {1099033B-8BD9-4038-AF9A-94A3370B4F8C}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1378F8F4-B354-4DB6-8052-A1A69B59BC2C}
+	EndGlobalSection
+	GlobalSection(MonoDevelopProperties) = preSolution
+		Policies = $0
+		$0.DotNetNamingPolicy = $1
+		$1.DirectoryNamespaceAssociation = PrefixedHierarchical
+		$0.TextStylePolicy = $4
 	EndGlobalSection
 EndGlobal
